rag_app/requirements.txt
streamlit>=1.31.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
pypdf>=4.0.0
python-docx>=1.1.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.18.0
